Test and Submission structs functional
The Test struct (could be renamed in future), contains information about
the nature of the correct result for a question - and thus how to
compile/run the Test.

The Submission is similar, but a little simpler. It contains information
about the format of the submission, e.g., is it going to be compiled
from sources, is it a compiled executable, etc.

Both contain as-of-yet unimplemented enums for addition of more
functionality down the line. Currently, both only use an executable
type, but Unit Test format should be made possible in future, and
compilation from sources.

Next target is to create a general Config for the Program, sourced from
a file. This will then be used to initialise TestRunner's in future, and
all other components of the program.