Configuration loading structure in place
The full configuration of the program is not yet well set up, but a
structure is in place to initialise config directories if they don't
exist, and a default configuration file.

At the moment, the plan for the structure is to have a modules/
directory in examtrainer's config, and an exam/ directory as well.

The exam/ directory will contain configuration files for different
exams, and these will have two different types (at least), named
"exam_type":
- 'specific': This will refer to specific modules, by name, that will be
  included in the exam. These modules must exist within the modules/
  directory
- 'general': This will contain an array of numbers, which refer to the
  difficulty of the module (or potentially, two values for a difficulty
  range). This will also require that modules have a 'difficulty' value.
  The order of questions in the exam will be in the order of the array,
  and a question with difficulty that lies in the relevant range will be
  picked randomly.
- 'all': This will create a random exam from every module that exists in
  the modules/ directory

There will also have to be an optional 'exam_order' key for
configuration:
- 'in_order': The default value, proceed with questions in the order
  they are listed.
- 'random': Ask the questions in a random order.

A lot of this is going to be difficult to manage, and it would make
sense to just implement one of these to begin with (although attempting
to keep the code flexible to allow for the addition of other methods in
the future).