Begun work on Exam struct
After an Exam struct is setup, I'll skip the step for now of choosing
between various exams and instead test to see if the program can be run
with a single exam, with a single question.

There's a lot of work to be done in a lot of different areas at the
moment.