A lot of progress towards pre-alpha
I added some functionality to make the output of the program resemble
the examshell, and some main function logic to control the program.

As I was reaching the stage to handle loading of modules however, I'm
quickly realising that the Exam structure being directly loaded from
Toml is not going to suffice, and that I'm going to have to use
something similar to the TestRunner, which reads Toml, but then applies
a lot of validation procedures to produce a structure that is directly
applicable to the program. I'm going to have to do the same with the
Exam.