Improved Exam structures, and program output
Almost ready to attempt running with an actual functional test.
To begin with, this will be a test with a single, specified test.

After this, the goal will be to run with a series of specified tests.

Running a General exam, which has a random element to its selection,
will require more modules, and more code.