About to implement grading function
There will be a lot more work to do in this area, but it would mark a
significant milestone if I'm able to run the program for 1 specific
exercise, and have it graded correctly.

The program is almost there, the TestRunner is set up, the Assignment
containing a TestRunner is located in an appropriate place inside the
Status struct, it just needs a few more hours and it's there.

After this, goals should be:

- Make sure the program can run for multiple exercises, correctly. This
  will form the groundwork for the program.
- Implement new evaluation mechanisms, allowing submission of an
  executable, or UnitTests, etc.
- Set up the Subject system appropriately.
- Implement a "difficulty" rating for modules, and then possibly the
  most difficult step, implement ExamType::General to work correctly
  using these difficulty levels - choosing appropriate questions
  randomly.