Almost at extremely basic prototype
Fixed a little bit of the logic around grading.

Prepared for the next stage after successful tests are run: to take the
result of the test and determine what the next assignment is going to
be. It is likely that it will not be essential to re-assign, but just to
adjust some of the values for a new assignment (new attempt number,
possibly less points, etc.)

The only step left to reach basic prototype, is to create the commands
to compile and test the code. The testing logic following binary
creation is mostly handled, I just need to build the command to compile
all sources, and to decide where it should be compiled to (although this
really doesn't matter too much, it should just be deleted immediately
afterwards anyway).