Examtrainer does work with new Exam & question
The examtrainer works well with new exams, and new questions, but there
are of course conditions in which it breaks, and the code is in no
standard to continue yet.

Next step is a major overhaul and redesign of much of the code
structure.

Future steps can follow on from that.