Pictures, exams and modules directories added
This commit is going to serve as an example for a "proof of concept", a
program that is successful, but not yet made to an appropriate standard.

From this point, another repository I've been working in is going to
become the 'develop' branch of this repository. This way, the MVP
demonstration can be seen by others on GitHub, alongside the
work-in-progress in the more complete program.