Added tests, changed design of Test a little
The previous implementation of the code would require Config for tests,
and a lot of the testing would have to take place inside of the
TestRunner instead of tests taking place in Test - due to the full path
of the executable or source files not being known by Test, and being
determined at the grade moment.

Although storing the full path may be more memory-intensive, it makes
many operations much simpler, and the memory cost is extremely minimal.

Creating a potential optimization for the future, but if it is to be
implemented it will be much easier to add in with a working program.