QuestionDB added
QuestionDB will serve as the interface between the program and it's
stored Questions. So far only the creation has been implemented, with a
`new` function. Some Documentation describing the QuestionDB has also
been created.

Error messages for invalid questions are displayed on the screen as
warnings during generation.

More extensive testing will be created when more of an API has been
implemented for the QuestionDB.