New Range class created
Range is to be used in the Exam toml, when creating a General Exam. It
will denote the range of difficulty values that a question at a certain
level should have.

The next task is to create a function for QuestionDB that will return a
vector of references to questions that fit within a given range.

A new test question has also been added for testing: aff_a, a question
that is in the real examshell.

In future, all questions should ideally be added to examtrainer, but
this is some distance into the future.