Added example submission
An example submission `hello_world` has been added to the tst/resources
directory. ProgramOutput has also had fmt::Display implemented, so it
will now be easy to convert it to a String, before writing it to a
Trace.

Next task is to create the Trace struct, before beginning tests with
Test & Submission.