Basic Trace struct added.
When a Test is run, it will return either a TestResult::Passed, or a
TestResult::Failed(Trace), with the Trace containing a string with the
reason or reasons for the test failure.

Trace output is well designed for when comparing an executable against
another - it may be appropriate however to make Trace into an enum
instead in future, to provide differently formatted traces depending on
the Test type.