Preparing for Test execution
And of course, the Unit Tests associated with Test.

To improve the code, and ensure maintainability, compilation
instructions have been abstracted to a new Compiler struct.
The idea behind it is to make compilation consistent across different
test types, including both those where the submission sources and test
sources are compiled together, and those where they are compiled
separately.

Additionally, it hides the logic for generating binary names, which are
generated randomly.