Tests planned in last commit message added
Question Database works well, and grading is now possible for ONE of the
possible Test types.

New tests run failing tests, and check the Trace output.

Minor note: I'm not completely satisfied with the form of the Trace
structure, mainly because it makes testing somewhat difficult (if the
Display of Trace changes, which would be a very reasonable change, then
many tests need to be changed as they rely on the to_string
function). If I ever feel the need to change the Display of Trace, then
I'll go further at that time and change it's internal structure. The
only reason I didn't do this first time is that the Trace structure will
vary depending on the type of Test, and it may be necessary to make
Trace into an enum also.