Improved Binary testing
I realized yesterday that with the current implementation for testing a
submission, the program would freeze while grading a submission that
contained an infinite loop, or similar.

Fixed by expanding the function that runs binaries, and collects output,
to limit the amount of time that a submission can run for.

In future, maybe this limit should be controlled by the Question. For
the time being, I'm satisfied with this implementation.