Config overhaul completed
A new more sophisticated Config has been created, with initialization
using the same Config struct as before, but now generating a struct that
is not simply created directly from TOML.

This new struct also has support for adding framework flags, in
preparation for the addition of unit tests. This will be the next step.