Validation added for test frameworks
It would be nice to prevent an invalid Test being created in the first
place, and at some point changes may be made to make this so, but it
would require Config to be passed along as a parameter.

For now, this is pretty good. The framework check is inside of
Question's constructors, so it is still impossible to have an invalid
framework within a Question.