Unit test implementation is in place
After a good day of effort - it's there. Using an external framework
(provided it is listed in the Config) is now possible for Unit tests.

There are of course additional checks that could be in place. For
example, some sort of function to validate the framework's
functionality. Right now, all that the framework validation does is:
- Check that all flags begin with '-l' or '-L'.
- If the flag is '-L', it checks that the path does print to a path that
  exists, and that the path is in fact a directory.

What Config validation does NOT do, is it does not confirm that the
framework is in fact present at the path. I don't think this would be a
valuable use of time however - and that this sort of error is the user's
responsibility given correct installation.