Started work on Exam loading
A Level struct has been created. An exam will consist primarily of a
vector of Levels.

Implementation to create a Level looks well formed, but at the moment
there just isn't a large enough question database to sufficiently test
correct generation of Exams.

Because of this, I spent some time writing a short script:
create_question.sh. This will simply generate the most simple template
for a question, using the "expected-output" type.

Of course, many questions are not going to fit this template, but it
should suffice for the development stages of this program.