Currently working on "Single question mode"
Before going for the full examshell, I decided it would be nice to begin
with a "single question mode", where you are tasked with answering just
a single question.

This will come with it's own associated set of commands, but it should
be a nice warm-up for the actual examshell.