Implemented logic to quit during Single Question
In Single Question Mode, there is now simple questions upon entering the
'quit' command.
- First it asks for confirmation that the user does want to quit.
- Then, it asks whether they would like the submit directories and
  subject directories to be deleted.

I think that this question concerning deletion of directories should not
be asked if the test is passed, but upon exiting the program as a whole,
there should be a question on whether the entire submission directory
and subject directories should be removed.