Basic functionality achieved for single question
Single Question Mode looks to be functioning very well, if a question is
answered incorrectly, a trace will be generated that is appropriate (if
the user requests one).

If answered correctly, the question is graded well, and the user is
returned to the main menu.

There are going to be a lot of small issues to iron out I'm sure, and
the code could certainly be tidied a bit more, but to prepare an Alpha
release by Wednesday (When the current Pisciners at Codam will begin
studying C for the first time), I'm going to focus on getting the Exam
logic sorted out, and creating a viable exam for practice.