Setting up question selection during Exams
A Level is responsible for taking a reference to a User, and using that
to return a question from that level.

An Exam is responsible for taking a reference to a User, selecting the
correct level, and requesting that the Level returns a question.

I think the code could be a bit more secure, at the moment there's a
bit of juggling required to make sure that the order of the code in the
shell functions is correct, so that a User always has an active question
when it needs one. It should be more than good enough for an alpha
though, once everything is prepared.